refactor: Rename unable_to_compact to possible_progress
